+++ src/main/java/com/sweetmimike/perfectmobfarm/PerfectMobFarm.java
@@ -7,13 +7,9 @@ import com.sweetmimike.perfectmobfarm.config.CommonConfigs;
 import com.sweetmimike.perfectmobfarm.item.ItemManager;
 import com.sweetmimike.perfectmobfarm.screen.MenuManager;
 import com.sweetmimike.perfectmobfarm.screen.MobFarmScreen;
-import com.sweetmimike.perfectmobfarm.utils.NbtTagsName;
 import net.minecraft.client.gui.screens.MenuScreens;
 import net.minecraft.client.renderer.ItemBlockRenderTypes;
 import net.minecraft.client.renderer.RenderType;
-import net.minecraft.client.renderer.item.ItemProperties;
-import net.minecraft.nbt.CompoundTag;
-import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.item.CreativeModeTab;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.level.block.Blocks;
@@ -66,36 +62,12 @@ public class PerfectMobFarm {
     }
 
     private void clientSetup(final FMLClientSetupEvent event) {
-        MenuScreens.register(MenuManager.MOB_FARM_MENU.get(), MobFarmScreen::new);
-
-        ItemBlockRenderTypes.setRenderLayer(BlockManager.IRON_MOB_FARM.get(), RenderType.cutout());
-        ItemBlockRenderTypes.setRenderLayer(BlockManager.GOLD_MOB_FARM.get(), RenderType.cutout());
-        ItemBlockRenderTypes.setRenderLayer(BlockManager.DIAMOND_MOB_FARM.get(), RenderType.cutout());
-        ItemBlockRenderTypes.setRenderLayer(BlockManager.EMERALD_MOB_FARM.get(), RenderType.cutout());
-        ItemBlockRenderTypes.setRenderLayer(BlockManager.NETHERITE_MOB_FARM.get(), RenderType.cutout());
+        ClientSetup.setup(event);
     }
 
     private void setup(final FMLCommonSetupEvent event) {
         // some preinit code
         LOGGER.info("HELLO FROM PREINIT");
         LOGGER.info("DIRT BLOCK >> {}", Blocks.DIRT.getRegistryName());
-
-        event.enqueueWork(() -> {
-            ItemProperties.register(ItemManager.MOB_SHARD.get(), new ResourceLocation(PerfectMobFarm.MODID, "completed"), ((pStack, pLevel, pEntity, pSeed) -> {
-                CompoundTag nbtData = pStack.getTag();
-                if (nbtData != null && nbtData.contains(NbtTagsName.KILLED_COUNT)) {
-                    return nbtData.getInt(NbtTagsName.KILLED_COUNT) == CommonConfigs.MOB_SHARD_KILL_NEEDED.get() ? 1.0F : 0.0F;
-                }
-                return 0.0F;
-            }));
-
-            ItemProperties.register(ItemManager.ADVANCED_MOB_SHARD.get(), new ResourceLocation(PerfectMobFarm.MODID, "completed"), ((pStack, pLevel, pEntity, pSeed) -> {
-                CompoundTag nbtData = pStack.getTag();
-                if (nbtData != null && nbtData.contains(NbtTagsName.KILLED_COUNT)) {
-                    return nbtData.getInt(NbtTagsName.KILLED_COUNT) == CommonConfigs.MOB_SHARD_KILL_NEEDED.get() ? 1.0F : 0.0F;
-                }
-                return 0.0F;
-            }));
-        });
     }
 }