+++ src/main/java/com/sweetmimike/perfectmobfarm/screen/MobFarmScreen.java
@@ -8,10 +8,13 @@ import net.minecraft.client.renderer.GameRenderer;
 import net.minecraft.network.chat.Component;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.world.entity.player.Inventory;
+import net.minecraftforge.api.distmarker.Dist;
+import net.minecraftforge.api.distmarker.OnlyIn;
 
 /**
  * Mob farm screen
  */
+@OnlyIn(Dist.CLIENT)
 public class MobFarmScreen extends AbstractContainerScreen<MobFarmMenu> {
 
     private static final ResourceLocation TEXTURE = new ResourceLocation(PerfectMobFarm.MODID, "textures/gui/mob_farm_gui.png");